requirements.txt
chainlit
phidata
python-dotenv
anthropic
